requirements.txt
Flask
python-dotenv
psycopg2-binary
sqlalchemy
PyJWT==2.8.0
requests
bcrypt
pytest